* Threads disabled.  They don't seem to work with OSX. * Added BIO process that models bioturbation.  Presently, it is a simple diffusion model       but is really just a place holder for a more complicated model. * Reorg of sedflux/src.       - Process header files are now in a single header file.       - Added destroy functions for each process to match the new Sed_process. * Sedlux is an ESMF grid component.       - Created init, run, and destroy functions for sedflux.       - Although it is able to compile and run within the ESMF, it has does very little         and has not been heavily tested. * Altered sedflux parameter (epoch) file       + Parent/child processes          - Plume is now a parent for hypopycnal and hyperpycnal plumes          - Failure is a parent for debris flow and turbidity current * Added more sedflux tests       - subsidence          + 3d test          + 2d test       - inflow       - sakura * Updated INFLOW       - Reorganized sakura/src          + inflow_sedflux_interface.c :  Functions to interface with sedflux architecture          + inflow_io.c :  Functions controlling input/output to inflow * Updated SAKURA       - Reorganized sakura/src to match inflow.          + sakura_sedflux_interface.c :  Functions to interface with sedflux architecture          + sakura_io.c :  Functions controlling input/output to sakura       - Removed runsakura       - Updated input test files * Updated docs       - Visualizing output with MATLAB       - MFILE documentation          + lex file mtoc.l * libutils/src       - Reorg of libutils/src.  Split functions in utils.c up into separate files.          + eh_file_utils.c: utilities to open/test/read files and directories          + eh_status_bar.c: A simple status bar.  Uses threads so is unsupported on OSX until I figure             out why threads aren't working there.          + eh_str.c: c-style string utilities          + eh_sequence.c: Sequences of data          + eh_polygon.c: Create and manipulate polygons          + eh_io.c: Input/output utilities          + eh_thread_pool.c: (not used) A deprecated implementation of thread pools          + eh_misc.c: Leftover utilities          + eh_messages.h: General message reporting macros          + eh_macros.h: Various macros       - Better error reporting for dlm files       - Scan time ranges.  (ex. 0y->1000y or 10y->100y,1000y->2000y) * libsed/src       - Reorganized          + sed_signal.c : handle signals for sedflux (moved from sedflux/src)          + sed_hydrotrend.c : functions that read / write hydrotrend output files       - sed_cube.c :  changed the way grids of suspended sediment are stored.  They are now          attached to each Sed_riv through a g_dataset.          + sed_river_attach_susp_grid          + sed_river_dettach_susp_grid       - sed_column.c : Faster methods for adding/removing Sed_cells.          + stack_cell_loc functions append pointers to Sed_cells to a Sed_column rather than copying          + sed_column_extract_[top_cell_loc,cells_above,top_n_cells] remove pointers to Sed_cells             and put them in a NULL terminated array of Sed_cells.       - Better error reporting.       - Epoch and Process queues: This is mostly code that was moved from sedflux/src       - Epoch input record changed.  Specify active time as a duration (ex. 0y->1000y) rather than as          an epoch number and a duration.       - Sed_process          + no longer record the size of the user data.  It is now the user's responsibility to malloc/free            the user data.          + run_at_end functionality       - sed_sediment.c :  get/set viscosity of water       - Sed_hydro          + Added member 't' to record the start time of the hydro event          + Fixed-up buffering of hydro data          + Better error reporting * sedutils/src/sedflux-read-bathy       - Read a hydrotrend file       - Swap bute order of hydrotrend file       - Create buffer based on based on flood magnitude

+++ libutils/src/eh_key_file.h
@@ -67,10 +67,10 @@ gboolean*     eh_key_file_get_bool_values  ( Eh_key_file f           ,
 double        eh_key_file_get_dbl_value    ( Eh_key_file f ,
                                              const gchar* group_name ,
                                              const gchar* key );
-double*       eh_key_file_get_dbl_array( Eh_key_file f           ,
+double*       eh_key_file_get_dbl_array( Eh_key_file      f          ,
                                              const gchar* group_name ,
                                              const gchar* key        ,
-                                             gssize* len );
+                                             gint*        len );
 double*       eh_key_file_get_dbl_values   ( Eh_key_file f           ,
                                              const gchar* group_name ,
                                              const gchar* key );
